requirements.txt
fastmcp>=2.12.0
nvm
python-jose[cryptography]
requests
python-dotenv
eunomia-mcp
black